with loggers

+++ OnlineMobileStore/src/main/java/com/mobile/application/controller/AdminController.java
@@ -3,6 +3,8 @@ package com.mobile.application.controller;
 import java.util.Objects;
 
 import org.modelmapper.ModelMapper;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
@@ -10,6 +12,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -31,7 +34,7 @@ import com.mobile.application.model.Payment;
 import com.mobile.application.model.User;
 import com.mobile.application.repository.ItemRepository;
 import com.mobile.application.repository.PaymentRepository;
-import com.mobile.application.repository.UserRepository;
+import com.mobile.application.repository.UserServiceImpl;
 import com.mobile.application.service.ItemServicesAdmin;
 import com.mobile.application.service.UserServicesAdmin;
 
@@ -39,13 +42,14 @@ import com.mobile.application.service.UserServicesAdmin;
 @RequestMapping("/Admin")
 @ResponseBody
 public class AdminController {
+	
 	@Autowired
-	PaymentRepository paymentRepo;
+	PaymentRepository paymentRepository;
 	@Autowired
 	ItemServicesAdmin itemService;
 	@Autowired
 	UserServicesAdmin userService;
-	UserRepository userRepo;
+	UserServiceImpl userRepo;
 	@Autowired
 	ModelMapper modelMapper;
 	@Autowired
@@ -53,6 +57,7 @@ public class AdminController {
 	@Value("${item_size}")
 	private int size;
 	Image image = new Image();
+	Logger log = LoggerFactory.getLogger(AccessoriesController.class);
 
 	/**
 	 * Displays all Registered Users
@@ -68,15 +73,18 @@ public class AdminController {
 		if (Objects.isNull(pageNumber))
 			pageNumber = 0;
 		if (Objects.isNull(size))
-			size = 25;
+			size = 50;
 		if (Objects.isNull(sort))
 			sort = "id";
 		Page<User> allUsers = null;
 		Pageable pageable = PageRequest.of(pageNumber, size, Sort.by(sort).descending());
-		allUsers = userService.getAllUser(pageable);
+		allUsers = userService.getAllUser(pageable);	
 		if (Objects.isNull(allUsers)) {
+			log.error("error found in userList");
 			throw new ItemNotfoundException("No Users Found.!");
 		}
+		log.info("AdminController userList() response{}", allUsers);
+
 		return allUsers.map(user -> {
 			return modelMapper.map(user, UserDto.class);
 		});
@@ -96,15 +104,18 @@ public class AdminController {
 		if (Objects.isNull(pageNumber))
 			pageNumber = 0;
 		if (Objects.isNull(size))
-			size = 25;
+			size = 50;
 		if (Objects.isNull(sort))
 			sort = "paymentid";
 		Page<Payment> allOrders = null;
 		Pageable pageable = PageRequest.of(pageNumber, size, Sort.by(sort).descending());
-		allOrders = paymentRepo.findAll(pageable);
+		allOrders = paymentRepository.findAll(pageable);
 		if (Objects.isNull(allOrders)) {
+			log.error("error found in orderList");
 			throw new OrderNotfoundException("No Orders Found.!");
 		}
+		log.info("AdminController orderList() response{}", allOrders);
+
 		return allOrders.map(orders -> {
 			return modelMapper.map(orders, PaymentDto.class);
 		});
@@ -125,16 +136,18 @@ public class AdminController {
 		if (Objects.isNull(pageNumber))
 			pageNumber = 0;
 		if (Objects.isNull(size))
-			size = 25;
+			size = 50;
 		if (Objects.isNull(sort))
 			sort = "model";
 
 		Pageable pageable = PageRequest.of(pageNumber, size, Sort.by(sort));
 		products = itemService.getAllItems(pageable);
 		if (Objects.isNull(products)) {
-
+			log.error("error found in productList");
 			throw new ItemNotfoundException("No products Found.!");
 		}
+		log.info("AdminController productList() response{}", products);
+
 		return products.map(item -> {
 			return modelMapper.map(item, ItemDto.class);
 		});
@@ -159,7 +172,7 @@ public class AdminController {
 	 * @param users
 	 * @return
 	 */
-	@PostMapping("/delete/{model}")
+	@DeleteMapping("/delete/{model}")
 	public Page<ItemDto> deleteItem(@PathVariable("model") int model,
 			@RequestParam(value = "page", required = false) Integer pageNumber,
 			@RequestParam(value = "size", required = false) Integer size,
@@ -169,15 +182,22 @@ public class AdminController {
 		if (Objects.isNull(pageNumber))
 			pageNumber = 0;
 		if (Objects.isNull(size))
-			size = 25;
+			size = 50;
 		if (Objects.isNull(sort))
 			sort = "model";
+		if (Objects.isNull(model)) {
+			log.warn("Enter correct model number");
+			throw new ItemNotfoundException("No products Found.!");
+		}	
 		itemService.deleteItem(model);
 		Pageable pageable = PageRequest.of(pageNumber, size, Sort.by(sort));
 		products = itemService.getAllItems(pageable);
 		if (Objects.isNull(products)) {
+			log.error("error found in deleteItem");
 			throw new ItemNotfoundException("No products Found.!");
 		}
+		log.info("AdminController deleteItem() response{}", products);
+
 		return products.map(item -> {
 			return modelMapper.map(item, ItemDto.class);
 		});
@@ -201,17 +221,24 @@ public class AdminController {
 		if (Objects.isNull(pageNumber))
 			pageNumber = 0;
 		if (Objects.isNull(size))
-			size = 25;
+			size = 50;
 		if (Objects.isNull(sort))
 			sort = "model";
 
 		Item item = modelMapper.map(itemDto, Item.class);
+		if (Objects.isNull(item)) {
+			log.warn("Enter correct Item Details");
+			throw new ItemNotfoundException("No products Saved.!");
+		}	
 		itemService.saveItem(item);
 		Pageable pageable = PageRequest.of(pageNumber, size, Sort.by(sort));
 		products = itemService.getAllItems(pageable);
 		if (Objects.isNull(products)) {
+			log.error("error found in addNewProduct");
 			throw new ItemNotfoundException("No products Found.!");
 		}
+		log.info("AdminController addNewProduct() response{}", products);
+
 		return products.map(items -> {
 			return modelMapper.map(items, ItemDto.class);
 		});
@@ -235,7 +262,7 @@ public class AdminController {
 		if (Objects.isNull(pageNumber))
 			pageNumber = 0;
 		if (Objects.isNull(size))
-			size = 25;
+			size = 50;
 		if (Objects.isNull(sort))
 			sort = "model";
 		if (itemService.getItemByModel(model) != null) {
@@ -245,11 +272,15 @@ public class AdminController {
 			Pageable pageable = PageRequest.of(pageNumber, size, Sort.by(sort));
 			products = itemService.getAllItems(pageable);
 		} else {
+			log.warn("Enter correct Model Number");
 			throw new ItemNotfoundException("No products Found.!");
 		}
 		if (Objects.isNull(products)) {
+			log.error("error found in updateItem");
 			throw new ItemNotfoundException("product cannot Update.!");
 		}
+		log.info("AdminController updateItem() response{}", products);
+
 		return products.map(items -> {
 			return modelMapper.map(items, ItemDto.class);
 		});
