+++ OnlineMobileStore/src/main/java/com/mobile/application/controller/OrdersController.java
@@ -4,6 +4,8 @@ import java.util.List;
 import java.util.Objects;
 
 import org.modelmapper.ModelMapper;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
@@ -11,6 +13,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -28,7 +31,7 @@ import com.mobile.application.model.User;
 import com.mobile.application.repository.CartRepository;
 import com.mobile.application.repository.ItemRepository;
 import com.mobile.application.repository.OrdersRepository;
-import com.mobile.application.repository.UserRepository;
+import com.mobile.application.service.UserServiceImpl;
 
 @Controller
 @RequestMapping("/User")
@@ -40,13 +43,15 @@ public class OrdersController {
 	@Autowired
 	private CartRepository cartRepository;
 	@Autowired
-	private UserRepository userRepository;
+	private UserServiceImpl userService;
 	@Autowired
 	private ItemRepository itemRepository;
 	@Autowired
 	private ModelMapper modelMapper;
 	@Value("${item_size}")
 	private int size;
+	Logger log = LoggerFactory.getLogger(OrdersController.class);
+
 
 	/**
 	 * Saves Users Orders from cart
@@ -65,15 +70,16 @@ public class OrdersController {
 		if (Objects.isNull(pageNumber))
 			pageNumber = 0;
 		if (Objects.isNull(size))
-			size = 25;
+			size = 50;
 		if (Objects.isNull(sort))
 			sort = "cartid";
 
 		Page<Orders> orders = null;
-		User userList = userRepository.findById(id);
+		User userList = userService.findById(id);
 		String email = null;
 		email = userList.getEmail();
 		if (Objects.isNull(email)) {
+			log.warn("Enter proper user id");
 			throw new UserNotfoundException("email id: " + email);
 		}
 		Pageable pageable = PageRequest.of(pageNumber, size, Sort.by(sort).descending());
@@ -84,15 +90,18 @@ public class OrdersController {
 					iterate.getTotal(), "IN");
 			orderRepository.save(newOrder);
 			cartRepository.deleteById(iterate.getCartid());
-			Item item = itemRepository.findById(iterate.getModel()).get();
+			Item item = itemRepository.findByModel(iterate.getModel());
 			item.setQuantity_available(item.getQuantity_available() - iterate.getQuantity());
 			itemRepository.save(item);
 		}
 		Pageable ordersPageable = PageRequest.of(pageNumber, size, Sort.by("orderid").descending());
 		orders = orderRepository.findAllOrdersById(id, ordersPageable);
 		if (Objects.isNull(orders)) {
+			log.error("error found in saveOrders");
 			throw new OrderNotfoundException("No Items in your Cart to order");
 		}
+		log.info("OrdersController saveOrders() response{}", orders);
+
 		return orders.map(allOrders -> {
 			return modelMapper.map(allOrders,OrdersDto.class);
 		});
@@ -105,7 +114,7 @@ public class OrdersController {
 	 * @param user
 	 * @return
 	 */
-	@PostMapping("/removeOrder/{id}/{orderid}")
+	@DeleteMapping("/removeOrder/{id}/{orderid}")
 	public Page<OrdersDto> remove(@PathVariable Integer id, @PathVariable("orderid") int orderid,
 			@RequestParam(value = "page", required = false) Integer pageNumber,
 			@RequestParam(value = "size", required = false) Integer size,
@@ -113,22 +122,32 @@ public class OrdersController {
 		if (Objects.isNull(pageNumber))
 			pageNumber = 0;
 		if (Objects.isNull(size))
-			size = 25;
+			size = 50;
 		if (Objects.isNull(sort))
 			sort = "orderid";
-
+		if(Objects.isNull(id) && Objects.isNull(orderid))
+		{
+			log.warn("Enter correct Details");
+			throw new OrderNotfoundException("id ="+id+" or orderid="+orderid+" is incorect");
+		}
 		Page<Orders> orders = null;
 		Pageable pageable = PageRequest.of(pageNumber, size, Sort.by(sort).descending());
 		orderRepository.deleteById(orderid);
+		
 		orders = orderRepository.findAllOrdersById(id, pageable);
 		if (Objects.isNull(orders)) {
+			log.error("error found in remove orders");
 			throw new OrderNotfoundException("No Items in your Cart to order");
 		}
+		log.info("OrdersController remove() response{}", orders);
+
 		return orders.map(allOrders -> {
 			return modelMapper.map(allOrders,OrdersDto.class);
 		});
 	}
 
+	
+	
 	/**
 	 * All yet to Complete Orders Every Order
 	 * 
@@ -143,19 +162,23 @@ public class OrdersController {
 		if (Objects.isNull(pageNumber))
 			pageNumber = 0;
 		if (Objects.isNull(size))
-			size = 25;
+			size = 50;
 		if (Objects.isNull(sort))
 			sort = "orderid";
 
 		Page<Orders> newOrders = null;
 		if (Objects.isNull(id)) {
+			log.warn("Enter correct User id");
 			throw new UserNotfoundException("User id: " + id);
 		}
 		Pageable pageable = PageRequest.of(pageNumber, size, Sort.by(sort).descending());
 		newOrders = orderRepository.findAllOrdersById(id, pageable);
 		if (Objects.isNull(newOrders)) {
+			log.error("error found in get every unpaid orders");
 			throw new OrderNotfoundException("No Items in your Cart to order");
 		}
+		log.info("OrdersController getEveryOrder() response{}", newOrders);
+
 		return newOrders.map(allOrders -> {
 			return modelMapper.map(allOrders,OrdersDto.class);
 		});
